Fix/error fixed/profileIntro/수정폼 이전내용 반영
